move isClionAvailable
